feat: Add support for charging rule installation and removal
The code changes introduce new AVPs and attributes to support charging rule installation and removal in the `ReAuthRequest` and `CreditControlAnswer` classes. This includes the addition of the `charging_rule_install` and `charging_rule_remove` attributes, as well as the corresponding grouped AVP classes `ChargingRuleInstall` and `ChargingRuleRemove`. These changes enable the implementation of charging rule-related functionality in the application.

Note: This commit message has been generated based on the provided code changes and recent repository commits.