refactor: Update Node class to include firmware revision
The code changes in `node.py` update the `Node` class to include the firmware revision attribute. This is achieved by adding the line `msg.firmware_revision = 240000` in the `send_message` method. This change allows the application to send the firmware revision information in the Diameter messages.

Note: This commit message has been generated based on the provided code changes and recent repository commits.